+++ src/lib/mapCreator.h
@@ -13,18 +13,6 @@
 
 const int MAP_TALLNESS = 25;
 
-std::vector<NodePtr> findLeafNodes(const NodePtr& rootNode);
-
-class MapGenerator {
- public:
-  std::vector<NodePtr> allNodesCollection;
-  int mapWidth, mapHeight;
-
-  MapGenerator(int mapWidth, int mapHeight);
-
-  std::vector<NodePtr> calculateMap(int maxIterations, int minRoomWidth, int minRoomHeight);
-};
-
 class MapCreator {
  public:
   int mapWidth, mapHeight;
